requirements.txt
numpy<2.0.0
streamlit
tensorflow
ultralytics
opencv-python-headless
numpy
pillow